Test if ai reviewer catches incorrect refactoring
